+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 ## 0.0.2 Second release (Android Support) [15 feb 2021]
 ## 0.0.3 Third release (Bug Fixes -> Platform issue)[16 Feb 2021]
 ## 0.0.4 Third release (Bug Fixes -> Platform issue)[16 April 2021]
+## 1.1.0 Fourth release(Bug Fixes-> Killed state issue )[17 August 2021]
+* Release
+   * Support Mediation
+   * Support Custom Notification Sound 
+    * Handle the range version sdk
+    * Bug Fixes
+    
 
-
-* TODO: Describe initial release.